Initial release
The bot is able to reply to an user interaction, with either a
random answer from a list or with a more customized action.

It's also able to reply multiple answers to one single message.

Answers every message ending with a question mark ("?") with a random
answer (yes, no, maybe, or evades the question)