Improve handling of timezones
If the user gives an actual timezone like Europe/Stockholm, we save this
into a table for use later. When somebody runs a remind command, if we
have the actual timezone saved we can then produce a more accurate
offset.

Note that this will always consider the delta from the current point in
time. Further improvement might be to look at when the reminder is
taking place (when a fixed point in time is given), and use the offset
at that point in time.

This change should be backwards compatible, for users to get the
improved timezone handling, they'll have to use the /tz command again to
insert their timezone into the new table. Users who don't, or who give a
time instead of a timezone, will get the old behaviour.

This updates LonamiWebs/Lonabot#10.